# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/users": {
      "get": {
        "summary": "Get Users",
        "description": "Get all users (Paginated).",
        "operationId": "get_users_users_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 12,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/roster": {
      "get": {
        "summary": "Get User Roster",
        "description": "Get user roster for team building (Public Safe Data).",
        "operationId": "get_user_roster_users_roster_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UserCandidate"
                  },
                  "type": "array",
                  "title": "Response Get User Roster Users Roster Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/users/{user_id}": {
      "get": {
        "summary": "Get User By Id",
        "description": "Get user by ID (Self or Admin only).",
        "operationId": "get_user_by_id_users__user_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/{user_id}/skills": {
      "get": {
        "summary": "Get User Skills",
        "description": "Get user's skills (Self or Admin only)",
        "operationId": "get_user_skills_users__user_id__skills_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update User Skills",
        "description": "Update user's skills (Self only)",
        "operationId": "update_user_skills_users__user_id__skills_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSkillsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/submit-assessment": {
      "post": {
        "summary": "Submit Assessment",
        "operationId": "submit_assessment_submit_assessment_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OceanSubmission"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfile"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/{user_id}/analysis": {
      "get": {
        "summary": "Get User Analysis",
        "description": "Get AI Analysis (Self or Admin only)",
        "operationId": "get_user_analysis_users__user_id__analysis_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/me/assessment": {
      "post": {
        "summary": "Submit My Assessment",
        "operationId": "submit_my_assessment_users_me_assessment_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OceanSubmission"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfile"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OceanSubmission": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "openness": {
            "type": "integer",
            "title": "Openness"
          },
          "conscientiousness": {
            "type": "integer",
            "title": "Conscientiousness"
          },
          "extraversion": {
            "type": "integer",
            "title": "Extraversion"
          },
          "agreeableness": {
            "type": "integer",
            "title": "Agreeableness"
          },
          "neuroticism": {
            "type": "integer",
            "title": "Neuroticism"
          }
        },
        "type": "object",
        "required": [
          "openness",
          "conscientiousness",
          "extraversion",
          "agreeableness",
          "neuroticism"
        ],
        "title": "OceanSubmission"
      },
...
      "SkillItem": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "level": {
            "type": "integer",
            "title": "Level"
          }
        },
        "type": "object",
        "required": [
          "name",
          "level"
        ],
        "title": "SkillItem"
      },
...
      "UpdateSkillsRequest": {
        "properties": {
          "skills": {
            "items": {
              "$ref": "#/components/schemas/SkillItem"
            },
            "type": "array",
            "title": "Skills"
          }
        },
        "type": "object",
        "required": [
          "skills"
        ],
        "title": "UpdateSkillsRequest"
      },
...
      "UserCandidate": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "character_class": {
            "type": "string",
            "title": "Character Class"
          },
          "level": {
            "type": "integer",
            "title": "Level"
          },
          "is_available": {
            "type": "boolean",
            "title": "Is Available"
          },
          "skills": {
            "items": {
              "$ref": "#/components/schemas/SkillItem"
            },
            "type": "array",
            "title": "Skills",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "character_class",
          "level",
          "is_available"
        ],
        "title": "UserCandidate"
      },
      "UserListResponse": {
        "properties": {
          "users": {
            "items": {
              "$ref": "#/components/schemas/UserPublic"
            },
            "type": "array",
            "title": "Users"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "users",
          "total"
        ],
        "title": "UserListResponse"
      },
      "UserProfile": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "character_class": {
            "type": "string",
            "title": "Character Class"
          },
          "level": {
            "type": "integer",
            "title": "Level"
          },
          "ocean_scores": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Ocean Scores"
          },
          "access_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Access Token"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "character_class",
          "level",
          "ocean_scores"
        ],
        "title": "UserProfile"
      },
      "UserPublic": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "character_class": {
            "type": "string",
            "title": "Character Class"
          },
          "level": {
            "type": "integer",
            "title": "Level"
          },
          "ocean_openness": {
            "type": "integer",
            "title": "Ocean Openness"
          },
          "ocean_conscientiousness": {
            "type": "integer",
            "title": "Ocean Conscientiousness"
          },
          "ocean_extraversion": {
            "type": "integer",
            "title": "Ocean Extraversion"
          },
          "ocean_agreeableness": {
            "type": "integer",
            "title": "Ocean Agreeableness"
          },
          "ocean_neuroticism": {
            "type": "integer",
            "title": "Ocean Neuroticism"
          },
          "is_available": {
            "type": "boolean",
            "title": "Is Available"
          },
          "skills": {
            "items": {
              "$ref": "#/components/schemas/SkillItem"
            },
            "type": "array",
            "title": "Skills",
            "default": []
          },
          "active_project_end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Project End Date"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "role",
          "character_class",
          "level",
          "ocean_openness",
          "ocean_conscientiousness",
          "ocean_extraversion",
          "ocean_agreeableness",
          "ocean_neuroticism",
          "is_available"
        ],
        "title": "UserPublic"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}